+++ changelog.md
@@ -2,7 +2,9 @@
 This changelog is based on https://keepachangelog.com/en/1.0.0/.
 
 ## [Unreleased]
-
+- Add automatic binary featurizer based on EBM
+- Add base class for score card models
+- Add Optuna based score card model
 
 ## [0.1.0] - 2022/11/15
 ### Added